feat: time range slider moved at the top #71
